% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{osteoHIRA1223-package}
\alias{osteoHIRA1223}
\alias{osteoHIRA1223-package}
\title{osteoHIRA1223: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
